requirements.txt
Flask==3.0.0
Flask-Session==0.4.0
selenium==4.17.2
webdriver-manager==4.0.1
pymongo==4.6.1
google-auth-oauthlib==1.0.0
pyotp==2.8.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow==10.0.0
requests==2.31.0 